+++ hw05/src/main/java/yaklyushkin_otus/spring_2019_02/hw05_app/repository/AuthorRepository.java
@@ -24,7 +24,7 @@ public class AuthorRepository implements AuthorDAO {
 
     @Override
     public Author insert(Author author) {
-        int authorId = this.getMaxId();
+        long authorId = this.getMaxId();
         author = new Author(authorId,
                 author.getAuthorSurname(), author.getAuthorName(), author.getAuthorPatronymic());
         final Map<String, Object> parameters = this.construct(author);
@@ -55,13 +55,13 @@ public class AuthorRepository implements AuthorDAO {
     }
 
     @Override
-    public Author deleteById(int authorId) {
+    public Author deleteById(long authorId) {
         final Author author = this.getById(authorId);
         return this.delete(author);
     }
 
     @Override
-    public Author getById(int authorId) {
+    public Author getById(long authorId) {
         final Map<String, Object> parameters = new HashMap<>(1);
         parameters.put("author_id", authorId);
         return this.jdbcTemplate.queryForObject(
@@ -85,7 +85,7 @@ public class AuthorRepository implements AuthorDAO {
 
         @Override
         public Author mapRow(ResultSet resultSet, int i) throws SQLException {
-            final int authorId = resultSet.getInt("author_id");
+            final long authorId = resultSet.getLong("author_id");
             final String authorSurname = resultSet.getString("author_surname");
             final String authorName = resultSet.getString("author_name");
             final String authorPatronymic = resultSet.getString("author_patronymic");
@@ -93,9 +93,9 @@ public class AuthorRepository implements AuthorDAO {
         }
     }
 
-    private int getMaxId() {
-        Integer maxId = this.jdbcTemplate.queryForObject(
-                "select max(author_id) from authors", this.emptyMap, Integer.class);
+    private long getMaxId() {
+        Long maxId = this.jdbcTemplate.queryForObject(
+                "select max(author_id) from authors", this.emptyMap, Long.class);
         if (maxId == null) {
             return 1;
         }
